refactor(wizards/subnetwork): move wizards
