+++ test/unit/editors/communication/SubNetwork.test.ts
@@ -2,15 +2,15 @@ import { fixture, html, expect } from '@open-wc/testing';
 
 import '../../../../src/wizard-textfield.js';
 import {
-  createSubNetworkAction,
-  updateSubNetworkAction,
-} from '../../../../src/editors/communication/subnetwork-editor.js';
-import {
   WizardInput,
   isCreate,
   isUpdate,
   isDelete,
 } from '../../../../src/foundation.js';
+import {
+  createSubNetworkAction,
+  updateSubNetworkAction,
+} from '../../../../src/wizards/subnetwork.js';
 
 describe('SubNetworkEditor', () => {
   describe('with no nulled properties', () => {
